+++ src/logger/index.ts
@@ -6,15 +6,17 @@ import { wLogWithSuccess } from "../utils/log"
  * @param props see interface LoggerOptions from winston
  * @returns winston instance
  */
-export const useLogger = (options?: LoggerOptions) => {
+export const withLogger = (options?: LoggerOptions) => {
   wLogWithSuccess(`[Wertik-WinstonLogger]`, `Initialized winston logger`)
   return winston.createLogger(options)
 }
 /**
  * Allows creating multiple logger instances
- * @param fn callback function, useWinstonTransport expects a function and useWinstonTransport runs that function with winston passed so you can return transport instances
+ * @param fn callback function, withWinstonTransport expects a function and withWinstonTransport runs that function with winston passed so you can return transport instances
  * @returns should return array of winston transport object.
  */
-export const useWinstonTransport = (fn = (winstonInstance = winston) => []) => {
+export const withWinstonTransport = (
+  fn = (winstonInstance = winston) => []
+) => {
   return fn(winston)
 }